Add command to create consolidated AIO config
With the AIO configuration spread amongst several different files, it
may be hard to see how they fit together. This command provides a way to
generate a single, combined configuration to see what it would look like
prior to being fed to the dynamic inventory.

Also, it can serve as a starting point for projects working on OSA
integration.

A new script, gen-config.py, is introduced to expose this functionality
to end users.

Parent-Id: I577cdbf4aadfcce846412edd7e2a394c257c0243
Change-Id: I4ffa621e72483957bbce0057b05a9e5b33d27b7c