[docs] Update branch names for Ocata
This patch adjust the branch names for Ocata to ensure that
the install guide and upgrade guide references are correct.

Also, the word TAG is replaced with the substitution reference
to ensure that the latest tag is always shown once the docs
move into a stable branch.

Finally, the developer note regarding the supported platform
for an AIO has been updated to reflect the Newton platform
support.

Change-Id: I1a8464917343c9533fe5698ab8da18bf1e1529d7