Set the AIO to use bootstrap_host_public_interface
This patch namespaces the variables 'public_interface' and
'encapsulation_interface' to prevent any clashes, and sets
the AIO networking to configure the masquerade rules to use
'bootstrap_host_public_interface' as the interface it uses
instead of hard-coding it to eth0.

This allows testing with hosts which don't have eth0.

Change-Id: I9a494ae5290a2218cf8c8d2cd44b27fd7264313f