Standarize usage of format strings
Previously our code was using both the older %-style and format-style
interpolation methods. This change consolidates the code base such that
the style used is consistent.

Logging lines were not changed, due to the logging library not
supporting format strings. The variables to interpolate are passed as
arguments and not actually interpolated unless the logging level is set;
this is a performance gain over making all strings passed to logging
methods pre-formatted.

Change-Id: I4fb7b821dd6f5662ef30ef67c456849a1d92da68