Use the correct upper constraints file
The repo generated file requirements_constraints.txt contains
git source references. This is useful for the wheel build
process, but not when trying to use the file as a remote
constraint file.

If using this file, the remote install process will try to
install, for example:
git+file:///var/www/repo/openstackgit/aodh@62b18d667064506b3f8f62653fc0b8662b41ebf5#egg=aodh

This would only work on the repo server because that's where
the git clone is held and where that file reference is.

In order to make a pip install work using a repo-hosted
constraints file, the right file to use is:
requirements_absolute_requirements.txt

Its reference looks more like 'aodh==1.1.3.dev1' which is
the exact version for the wheel at the above-mentioned SHA.

Change-Id: I44b8eab8f36f044932d2dc9ca0d5013f0f17d38b