Implement ARA logging and reporting for openstack-ansible gate jobs
ARA[1] is a tool that records ansible-playbook runs and makes the
data available over CLI or through a web application that can also
be statically generated.
It's purpose is to help visualize and understand playbook runs in
order to make them easier to troubleshoot.

[1]: https://ara.readthedocs.io/en/latest/faq.html#what-is-ara

Change-Id: Ie364212a915d51713e85b75c13ace42674840561
(cherry picked from commit 1053adfa77c142175833bdeade181564afae84b0)