Fix multi arch repo-build
The existing multi arch repo build is broken because the add_host module
bypasses the host play loop of ansible and is only executed on the first
host. To excute on multiple hosts the 'with_' directive must be used.
The result of the existing code is that the repo_nodes group is always
empty because the first host is always skipped by the when clause.

Change-Id: I2c37217a78c006ae28a4f31f3d86244ef4df60be
Closes-Bug: #1641410
(cherry picked from commit 1e0c11f4165d8e74c028e5e439b93ca083e1515d)