Consolidate tox run output into one report
The original implementation of erasing the coverage data file between
each test was the most accurate method, providing coverage stats in a
'cleanroom' setup, but resulted in extra reports being output. This
doesn't scale very well as more test files are added to the repository.

This patchset removes the extranenous erase command, and adds the `-a`
argument to the run commands, which allows coverage to accumulate data
across runs. Without the `-a` run, the coverage reports for each target
file are cross-contaminated, resulting in different coverage numbers.

Depends-On: I06729ac2bc1688a39255f2c8ea0d14131b5c2560

Change-Id: I4d52f7363e00bba62ed878fe8fa6e593eb21fdde