DynamicInventory filesystem storage API cleanup
Update some function docstrings for accuracy/completeness
Make some methods module-private as they are not required to be
available for external use.

Tests were updated to accomodate for the now private functions.
The affected tests were identified as filesystem-centric and
moved to their own testing file, which will run under the inventory
tox target as well.

Parent-Id: I4d52f7363e00bba62ed878fe8fa6e593eb21fdde
Change-Id: I0275eeeafeb2a4c89cd820d56dd0ad01d3124a80