Inventory filesystem refactor load funcs moved
The last uses of the builtin os module are removed from generate
as the load_environment and load_user_configuration functions are
moved, along with the _extra_config method they rely upon. This
also moves the last use of the yaml module.

These functions are not refactored to make use of the other funcs
available in the filesystem package, and as a result there is
still an undesirable use of the dir_find func in the generate
script. These will all be cleaned up in a following patch to keep
this easier to review.

An unused left-over function is removed now that there are no
merge conflicts introduced by it.

A debug logging message is formatted consistent to other messages;
the only impact in this case is applying consistency in log
message formatting since the formatting was only occuring when the
log level is activated already.

Parent-Id: I4c7b45f16f8ffecb91bf509daeca47b2c2450681
Change-Id: I577cdbf4aadfcce846412edd7e2a394c257c0243