Use get-pip script from repo release folder
This patch ensures that if the local repo server has a copy of the
get-pip.py script then it will be used instead of reaching out to
the default location on the internet.

The tasks to set upper constraints are removed form the unbound
and repo-server playbooks as those roles do not execute any form
of pip installation.

This approach is taken to ensure that:

- Before the repo is built, the get-pip process in the pip_install
  role still has two chances to fetch the get-pip.py script. If we
  simply override the first URL option in group_vars then the
  initial task will fail before the repo is built for the release
  being executed and we would therefore only get one fallback.

- After the repo is built, if the lxc-host-setup and other pre-repo
  playbooks are run, the plays will use the repo's get-pip script
  instead of reaching upstream to fetch a fresh one. This maintains
  a consistent get-pip script version within the same tag forever.
  If group_vars are set for the targets of these plays to use the
  upstream URL then the experience is not the same - the lxc_hosts
  group will always get the latest upstream get-pip script every
  time the playbook executes, resulting in the lxc hosts having a
  different copy of the script to other hosts and containers.

Partial-Bug: #1633105
Depends-On: I3735ec50d51e10d0bc4a7f6677a800444fb33912
Depends-On: I4a2ff6c6e5cd40d87d4218f63c09e96da82da1aa
Change-Id: I9b558a14e800bb77d78021f665173dcdfaf2f4aa
(cherry picked from commit 47d3f8733f76d567d2f578be20894bdfc5a6246c)