Bump horizon SHA to 10.0.0.0b1
The version of horizon currently included in openstack_services.yml is
broken.

This patch advances the SHA to 10.0.0.0b1 which is in a better state.

I'm aware that this may not be the proper process to bump the SHA,
but want to get this out so that we can get a horizon we can login to
included in the current AIOs.

Change-Id: I25dbb6e700e2249717b21b7026f95f6bf70a1721