Fix cinder_volumes is_metal override example
The example file that shows how to use an env.d file to override
the is_metal value for a service is functionally incorrect.

The file is missing the properties: dictionary level and without this
correct nesting of dictionaries, the is_metal value specified has no
effect on the inventory.

Change-Id: If73d22ff7dd6e5fec24f79217f721797effab651