Retry the git clone in role fetcher
I see failures in my local CI 10-20% of the time on cloning one or
two roles, so the test fails to bootstrap. I'm not sure if there
is some rate limiting in place on the Openstack git repos, but it
seems to fail quite often on a very reliable link.

The failure being seen is:
Failed to connect to git.openstack.org port 443: No route to host

This patch adds retries to the git clone task used to fetch roles
to hopefully avoid this failure.

Change-Id: I1ea74e3ac4b2d0da88f3155d593eaceb0a205d46
(cherry picked from commit c8606c443c9cbf77043f763973abebc6d7bb3791)