Fix default IPv6 BGP peer for BIRD
The jinja calculation of the BIRD BGP peer was failing intermittently
when IPv6 connectivity was not present.

Even though the IPv6 deployment was being detected correctly as
disabled by the role, the configuration task was failing to skip
consistently because the IPv6 peer address var could not be
resolved by Jinja.

This fixes the default IPv6 peer calculation to resolve to an empty
string when no IPv6 connectivity is present, which allows the
automatic IPv6 detection to skip the config task.

Change-Id: If450b237ba349a163f44f40b945c98b14c9d7042