Don't delete container_cidr key when overriding
Commit f426eb98d1e4dc0b8b65f9191b038edcea849755 added a global_override
deletion mechanism, but did not account for the automatically-created
container_cidr key.

This change ensures the key is skipped in deletion, and adds further
tests for the _parse_gloabl_variables function.

Backported to Newton manually because the dynamic_inventory.py file was split
at the beginning of the Ocata cycle.

Change-Id: I5c48adea5920c1524be75939a75af6bed3a9743a
Closes-Bug: #1649632