Call load_environment before mocking it
The load_environment call happened after mocking, resulting in a mock
object being used for the duplicated IP test. Inspecting
inventory.keys() during this test shows a length of 29, with only the
groups defined in the config file being present.

Moving the call outside the mocked context and providing data instead
moves the number of keys up to 240, which includes the entries from the
environment.

While this change isn't necessarily material to the check for duplicated
IPs, it is closer to the intent of exercising the generation system
fully.

Change-Id: I1e17e3ca4227f1da3fc7aa8accda4e93fa90d245