Update Add a new infrastructure host script part
Add "Replace <NEW INFRA NODE> with the name of the new host." to make
it clearer.

remove the file seperator after awk because if we use '|' as it,
the add_host.limit will have a white space at first col.

Add an new script to deal with the fatal error when we add the new
infra node. The 'Distribute self signed ssl key' playbook needs
horizon.pem from the first horizon container, so we have to add the
first horizon container name to the add_host.limit.

Change-Id: Ia6774a6e3ec8fe13dde7559942e2f899e18e8eb1