Ensure RHEL 6 STIG content is used
The RHEL 7 STIG content in the security role is ready to be enabled
as the default for deployers who consume only the security role itself.
However, the new content isn't ready for the integrated repo.

This patch ensures that RHEL 6 STIG content is used in the integrated
repo even if the security role itself switches to enable RHEL 7 content
as the default.

Change-Id: I1b7ceb90673e07c7fc5a2c882c84a56aa1d5877a