Pass correct venv paths to ceph_client
ceph_client was hard coded to use openstack_release as the venv tag.
Now it builds the venv path the same way the roles do to ensure the
vars match and the rbd libs are deployed to the correct venv path.

Also removes an extraneous var in the Gnocchi playbook that was
being passed to ceph_client but not used in the role.

Change-Id: I0a7bfe48bd86b8d7bc8dc344ba851c9493b00920