Fix fastest mirror fact set
Aside from the ansible bug workaround not working, an issue was
introduced by I5a018c88dde73564d15e15535d080adc5415e663 which
causes the fastest mirror fact set output to render incorrectly.

It should use stdout_lines instead of stdout, because running
the "list" filter on stdout (a string) causes the string to be
split into a char-by-char list of the string, rather than a
single element list containing the whole string.

This removes the call to the list filter and uses the list
in stdout_lines that already exists.

Unfortunately this does not work around the ansible bug but it
is cleaner than reverting to the original fact set so I would
prefer to fix the updated implementation instead of reverting
I5a018c88dde73564d15e15535d080adc5415e663.

Change-Id: Ic8eb28f83618f6caa8d5cf4094563f07e302a29f