Ensure that tox uses global pins
The tox implementation currently does not make
use of the global pins as requirements or as
constraints. This may cause an inconsistent
pip, setuptools and wheel package set being
installed when executing tests.

This patch ensures that the pinned versions are
installed when setting up the tox venv, and
ensures that the pins are used as constraints
in any subsequent package installations.

Closes-Bug: #1658773
Change-Id: Id4176f39eeefe3323c35e5f05bca6b208b615088