Add support for placement_database and cell_v2
This patch adds support for the cell0 database creation, and adjusts the
mysql-db-user.yml playbook to allow append_privs to be specified using
the "db_append_privs" var - defaulting to "omit".

The nova_placement_database is created as is the nova_cell0 database
which has permissions for the nova_api user.

Lastly there is a work around for bugfix:
https://review.openstack.org/#/c/428120/

Once that merges we can remove the "reserved_host_disk_mb: 0" setting.

Change-Id: I5fab3fbe982e12ca05adc051533f81da1093f7c0