Ensures the host IP isn't used for containers
We statically configure the host IP for many networks during the
bootstrap AIO process. Later, when the inventory is generated,
we use an IP from the queue, which could be conflicting with the
host (Probability: 1 on 204).

This should fix it.

Change-Id: I5a49eebef48892d7266684f1ffdbe1d08d82c211
(cherry picked from commit 96c766004de478f08026754e3d0443ebf8b6f467)