Target unbound component group instead of _all
The unbound container group and component group share the same name,
so we are hitting the quirk in OSA inventory where containerized
services that have component and container groups sharing the same
name get both the hosts and containers added to the _all group.

This change targets the component group to correctly target only
the unbound containers, not the physical hosts for the containers.

Change-Id: Id214bbebeee73a6bf48088396e2e450669e00926