SHA Bump for roles and upstream projects 15-02-2017
We should try keep as close to stable/ocata as possible before release
so when release comes it's a smaller increment.

This patch updates the SHA's to the latest stable/ocata role SHA and
head of stable/ocata for upstream projects.

Additionally, this patch reduces the processes for nova-placement uwsgi
app, and drops memory requirements for Galera to alleviate the OOM
errors in the gate.

Change-Id: Iada1e8d0392a528e0803215109167934b73f5313