[DOCS] Move limited connectivity section to Deploy Guide
The limited connectivity section was temporarily moved into
the developer guide temporarily when re-arranging the deploy
guide, but never moved back.

As this is important information for deployers to see, this
is being moved into the deploy guide appendix, then references
to it are added to the AIO and Deployment Guide in appropriate
places.

The following notes regarding additional changes apply:

- The pip offline install content for the limited connectivity
  page breaks the flow and doesn't really fit in the two models
  proposed. This content should move to the pip install role.

- The reference to the get_url/get_uri bug for Ansible 1.9 no
  longer applies as Newton onwards now uses Ansible 2.1.x and
  above.

- An unused Appendix H reference in the Security Appendix has
  been removed.

- The Appendices have been re-arranged slightly to try to show
  the information in a perceived order of importance.

Change-Id: If4b8a75277374ed7e96a1ce6610ed8a897125693