Cinder use correct glance API when RBD enabled
The syntax of the jinja block was broken and outputted a line break
always, like this:
ok: [lsn-d6329_cinder_volumes_container-9a2bbc7d] => {
    "msg": "True\n"
}

So the resulting evaluation was always false, and glance API v1 was
used by cinder-volume, breaking the ceph CoW functionality.

This fixes the variable to eval properly as a boolean and make
Cinder use the correct glance API version.

Change-Id: I9c1f31bba80bf2591710fd4d59f320ffadaf1c8b