Serialize Keystone and manage haproxy endpoints
Serialize keystone to run the playbook on 1 container first, then
run on the remainder of the containers in the environment.

This also adds a common-task include to disable the haproxy endpoints
for the containers where the role is running, then re-enable them
in the post tasks.

The haproxy endpoint management task will wait for up to 2 minutes
by default for the endpoint to report a healthy/UP or drained/MAINT
state.

Change-Id: I0df71255154cfb28b62bb14736ba0c8fc1f6ae66