Disable haproxy endpoint drain wait
When haproxy endpoint management has a wait condition for the
connection drain to occur, it causes the module to fail if a backend
is not specified. Since our current implementation relies on backend
auto discovery (since the same host may be present in multiple backends),
this causes the module to fail when it attempts to disable the endpoints.

Change-Id: Ia292e4cb4d8ba73a72af8305cc2006aea9c43e32