Move upper constraint logic to after python is installed
Currently if python is not already installed on the deployment host
the bootstrap-ansible script fails due to the way that the script
tries to check the python version installed (using python to do it).

This patch moves the logic until after python is actually installed
to cater for the situation where it may not already have been.

Change-Id: I06da3e84100b619e7c39962b7a43d4c8c7d707fa
(cherry picked from commit f438c65028abd7308b2b8ac85257a6a75ae5763d)