Remove ceilometer from the gate
This change removes ceilometer from the integrated gate. Sadly the
ceilomteter roles and playbooks have been a little neglected which has
resulted in a slow and sometimes unstable system. To improve gate times
ceilometer has been pulled out of the integrated gate. Once we're able
to get the ceilometer roles and playbooks  updated to the latest
release and we're able to resolve all of the deprecations we should
put these rolse back in rotation for the integrated gate.

Change-Id: I116928a1db96e242ed87177578581fe7bf16a001
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
(cherry picked from commit 5094011164a7a670a5c083269e50d82699eb9790)