Use an explicit version of urrlib3
If not forcing to install urllib3, ansible on the deploy node
will reuse the urrlib3 from site-packages, which could be older
than the one in global-requirements, because ansible doesn't
list it in its dependencies, so we don't explicit force a
certain version to be installed.

On my machine, I had an urllib3 installed with a version of
1.7.1, and all the lookups with https had SSL issues. Moving to
urllib3 fixed the thing.

This only cares about the ansible side, to ensure ansible venv
has the proper version. Hosts targetted by ansible (for example
when using get_url), also need a fix. This will be done in
different patches.

Change-Id: Ia07969ac9aee90724eecc6657affd1dbe16045a7