Adjust gate-check-commit for deploy/upgrade testing
In order to facilitate periodic deployment and upgrade testing,
this patch implements an 'action' as a second CLI parameter for
the gate-check-commit script.

This seperates the 'action' from the 'scenario', allowing us to
execute a gate-check-commit for multiple scenarios instead of
being stuck with only one. That is why the 'upgrade' scenario
is removed.

This patch is partnered with this project-config change:
https://review.openstack.org/419517

The patch also tidies a few things up in order to make the
script a little more readable.

Change-Id: Ie62efc188d5eafd5bb64eb53f14c191dd50bef33
(cherry picked from commit 2aa4a3f7d2bacfbbf58f6cf35f37ce654939dcff)