Use OpenStack-Infra MariaDB mirror in OpenStack-CI
When executing tests within OpenStack-CI, the MariaDB
mirror implemented in CopenStack-Infra will be used
instead of the Rackspace mirror.

This should speed up OpenStack-CI testing and be more
reliable as it eliminates external dependencies and
does not rely on a single mirror as a source.

For the sake of record, the MariaDB mirror was
implemented in OpenStack-Infra with these reviews:

- https://review.openstack.org/307831
- https://review.openstack.org/354203
- https://review.openstack.org/440769

Change-Id: I7254bc246525488f1b94a3835786307057a906c1
(cherry picked from commit daca9307bb7f8876a4a107a37a6b5225bcb758a8)