AIO: Provide facility to exclude apt distributions
When using a mirror which compiles all apt sources into
a single repo with the intent of relying on that repo for
all package updates the /etc/apt/sources.list file must
contain only the base ubuntu repo.

This patch adds the ability to implement an AIO with
this kind of apt sources configuration, while keeping
the current configuration that's used for gating.

Change-Id: Ia9785a43c9027df86f3bc94f825aba3c43a96e4c