[docs] Add http public endpoint example
This patch intends to expose how to use the same address
for both external and internal load balancer endpoints
while also covering the fact that it can't be done if
you're using different protocols for each endpoint.

This is to head off the common questions we get based
on mistaken configuraitons and misunderstandings.

Change-Id: I53d4e9751f5427e5bf6c2ab73ab3225f20789988