Fix upgrade script-library check
The script-library check is supposed to use a function
available in the library in order to validate that the
library has been loaded, and to fall back to loading
the library.

Currently this does not happen, so the library is never
loaded.

Change-Id: I8ce788448026fd26ade0c4e8b2895a0c89e62604
(cherry picked from commit 20b068ddaf65afaf8906e44d935cc1ce81471773)