Ensure that the terminal type is set
The terminal type is set to 'unknown' by default and is set
when a user logs in to a terminal. The 'unknown' terminal
type causes the 'clear' command in the upgrade job to fail.

We set the terminal type for the gate-check-commit script as
it is not using a tty.

Change-Id: Id460184dbfec1bc56bc05baa67924b5f724b5111