Introduce userspace group_vars and host_vars
This opens the door to using group_vars and host_vars in
userspace, by allowing the user to define a ":" delimited list
of folder that will then be used by a vars_plugin, and (maybe
later) by the inventory.

This can't override group_vars due to ansible precedence.
We could technically have this override done by moving the
group_vars into a different folder and adding it to these
environment variables.

Change-Id: Ib2cbb35c0a7bccd896fbe71c22eeb07188db99ee