Update all SHAs for 15.0.1
This patch updates all the roles to the latest available stable
SHA's, copies the release notes from the updated roles into the
integrated repo, updates all the OpenStack Service SHA's, and
updates the appropriate python requirements pins.

This is the first SHA bump for Ocata - so there is no dependent release
patch. Releases for Ocata will continue at a 2 week cadence along with
other stable releases from this point on.

Change-Id: I446e3113eff08c9653c859826460bf3a85146e42