SHA Bump on Master as at 22 March 2017
This includes changes to Tempest that were causing gate failures.
One further change is required - we need to consume the SHA of the
Depends-On patch (once it merges) into the nova-powervm repository,
before this patch will pass. This has been tested using a fork with
the fix, so it should work from that point on.

Finally - the HEAD request for Neutron returns a 405 now which causes
the haproxy VIP to fail. We need to move this to a GET request.

Change-Id: I68f6fef3ffdef3fbf3d215a95bf158dbe94f4b2c
Depends-On: I0b4818f3eb0b9bace9ee61e7f855fd070aee684b