Keystone: Source template files from git or deploy host
Instead of sourcing the paste/policy files from the role,
the deploy host will first be checked, then the git source.

This eliminates our need to carry the template files and
keep them up to date. It also ensures that if a custom git
source or alternative SHA is used, the correct source
templates are used.

Related-To: I910bb1700bafd48185c15d64401c8f8e93c696ea
Change-Id: I97476c42172cace5601f777e771ba0aa649b05ca