Converge neutron agents onto Baremetal
The neutron connectivity agents will be deployed onto baremetal along
side the defined "network_hosts". This gives deployers the ability to
simplify deployments, keep a standard topology, and scale to meet the
demands of the environment.

This addresses several long standing operational issues where the
neutron_agent container would be restarted, due to config changes,
upgrades, etc. This can cause a prolonged network outages when L3
tenant networks are in service. Moving these services to baremetal
resolves the issue permanently by ensuring consistent hardware
configurations, mac addresses, and simplifying the network transport
as no vethpairs or macvlans are in service.

Change-Id: Id16b1bcd67bf075a8677206a205ece010e7b014c
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>