Add variable for template filename
The current logic for constructing the template filename
does not allow custom scenario names. For example, if an
outside playbook defines ``boostrap_host_scenario`` to
be "swift" or "awesome_scenario", then the task would try
to look for the user_variables.swlft.yml.j2 template within the
role. This commit allows the user to define a template filename
using the variable bootstrap_user_variables_template.

Change-Id: I9fc98ea9b7d68623cb42c5fa3b9d5383fadb82d8