Use wait_for_connection instead of wait_for to check container
After creating the container or configuring the bind mounts for
the container, we used a task to wait for SSH to respond on the
container, using the Ansible wait_for module trying to connect to
port 22.

When managing containers in Ansible using other connection plugins
that use LXC commands on the physical host to execute modules,
there may not be an SSH daemon running within the container. In
these situations we can accomplish the same check without SSH by
using the wait_for_connection module (new in ansible 2.3) to verify
that Ansible has a working execution path into the container, but
not necessarily via SSH.

Change-Id: I1399b6727ec8894dd3b9c464f6addbeea1c1f4f9