Enable ARA cross-repo testing
When executing tests in OpenStack-CI, this patch
enables the use of 'Depends-On: <change-id>' in
the commit message to execute the test using a
patch that is still in review for the openstack/ara
repository.

This patch also lays the ground for similar testing
to be done with other openstack repositories, but
that will no work just yet as it will take quite a
bit more machinery to make the cloned repositories
be used to build the repo. That work will be
submitted in a subsequent patch.

Change-Id: Id946883fa14cd692fb7c6a0f24865b0239b7250c