Fix cinder backends detection for nova ceph deploy
When nova is not rbd backed but cinder is using ceph, the ceph_client
role should run against compute hosts. There is a var that is supposed
to resolve to a boolean indicating whether rbd backends are in use
on cinder, but it was broken and resolved to a string instead, breaking
the ceph_client conditional for nova.

Change-Id: I9999c373f942d3565ab86815e8330a489a83d677