[docs] Fix Newton deploy guide link
The current link goes to the index page for
all projects. This patch fixes it to point
directly at the OSA guide.

The link reference is also moved up into the
same location in the list as that used for
Ocata.

The protocol is also changed to HTTPS as all
docs websites are now published via HTTPS and
HTTP queries are redirected from HTTP to HTTPS.

Change-Id: I34b93adf12afc0c289b61c25a497314fc81c220e