Isolate ansible-runtime venv site packages
In the isolated ansible-runtime venv, we were inheriting the system
site packages inside the virtualenv. This breaks the openstack-ansible
wrapper because it always assumes that ansible is installed in the
ansible-runtime venv. When ansible is already installed on the system,
pip does not reinstall it in the venv because it is inherited from
the system, so the wrapper fails.

This fix also includes a PATH entry in the openstack-ansible wrapper
which will make the wrapper more resilient against bugs like this
in the future.

Change-Id: Ifa5234cf7353f1443c5b77787363d8880b5cf0b6
Closes-Bug: #1690447