scripts: Introduce sources-branch-updater-lib library
Move core functionality from sources-branch-updater to
sources-branch-updater-lib. The reason for that is that external
projects may also want to use a variant of the existing script so it's
best if we move the common functionality to a library which can be
sourced by OSA and external scripts.

External projects may use OpenStack Ansible to deploy their
infrastructure. They may use a well-known OpenStack branch for that
(ie stable/ocata) or work directly from master. The latter is common
in the external CI provided by OPNFV. In that case, the OSA master
might be too bleeding edge so we would like to have the ability
to pin down specific SHAs for roles and packages even when we are
deploying from master.

Change-Id: Ief1b2bc8338f0c11c87bea28a9cf37d4397ab4f1