Fix os_cmd script for tty
Add the '-1' parameter to lxc-ls so each container is listed on a
single line, otherwise multiple container names were being listed
on a single line and causing the command to fail when in a tty.

Also use the '--filter' parameter to return only the utility
containers rather than having lxc-ls process and return each
container.

Change-Id: Ifcf1f13392d17464a088ce2928f434387ca39365