Implement a data plane down time test for upgrades
This patch implements a VM using the demo project, then
consistently executes a connectivity test through L3
and a disk access test on the test instance while the
upgrade script executes.

The intent is to determine whether any data plane down
time is experienced while executing the upgrade.

Once the upgrade is complete, the testing is stopped
and if there are any failures, the job will exit as
failed.

Change-Id: Ie3f1738436c72eb82d7f6ca45d84024d11f3d923